chore(scripts): remove type declaration duplication within /scripts/monorepo and remove non public API
