+++ scripts/monorepo/index.d.ts
@@ -1,115 +1,11 @@
-export interface PackageJson {
-  name: string;
-  version: string;
-  private?: boolean;
-  main: string;
-  types?: string;
-  module?: string;
-  private?: boolean;
-  dependencies?: { [key: string]: string };
-  devDependencies?: { [key: string]: string };
-  peerDependencies?: { [key: string]: string };
-  [key: string]: any;
-}
-
-export interface PackageInfo {
-  /** Package path relative to the project root */
-  packagePath: string;
-  /** package.json contents */
-  packageJson: PackageJson;
-}
-
-export type AllPackageInfo = { [packageName: string]: PackageInfo };
-
-export declare function findGitRoot(): string;
-
-/**
- * Find all the dependencies (and their dependencies) within the repo for a specific package
- * (by default, in the CWD when this was called)
- */
-export declare function findRepoDeps(options?: {
-  /** Optional different cwd */
-  cwd?: string;
-  /** Whether to include dev deps (default true) */
-  dev?: boolean;
-}): PackageInfo[];
-
-export declare function getAllPackageInfo(): AllPackageInfo;
-
-/**
- * Determines whether a package is converged, based on its version.
- *
- * @param {Object} [options]
- * @param {PathOrPackageJson} [options.packagePathOrJson] - optional different package path to run in OR previously-read package.json
- * (defaults to reading package.json from `process.cwd()`)
- * @param {'library' | 'application' | 'all'} [options.projectType] - filter for what project types you wanna apply the condition
- *
- * @returns {boolean} true if it's a converged package (version >= 9)
- */
-
-export declare function isConvergedPackage(option?: {
-  packagePathOrJson?: string | PackageJson;
-  projectType?: 'library' | 'application' | 'all';
-}): boolean;
-
-/**
- * @param since - Commit to compare against
- * @returns Set of packages that are changed
- */
-export declare function getAffectedPackages(since?: string): Set<string>;
-
-/**
- * Returns SHA for the nth commit from a reference descending from the latest commit
- * @param n - nth commit from latest
- * @param ref - The github ref/branch the filter
- * @returns - A git commit SHA
- */
-export declare function getNthCommit(n = 1, ref = 'HEAD'): string;
-
-/**
- *
- * Gets project metadata from monorepo source of truth which is `workspace.json`
- */
-export declare function getProjectMetadata(options: {
-  /**
-   * repo root path
-   */
-  root?: string;
-  /**
-   * package name
-   */
-  name: string;
-}): import('@nrwl/devkit').ProjectConfiguration;
-
-/**
- * Returns all the dependencies of a given package name
- * @param packageName - including `@fluentui/` prefix
- * @param options
- * @param options.dev - include dev dependencies
- * @param options.production - include production dependencies
- */
-export async function getDependencies(packageName: string, options?: { production?: true; dev?: true }): string[];
-
-export function getLernaAliases(
-  options: {
-    excludedPackages: string[];
-    type: keyof typeof import('lerna-alias');
-  } & import('lerna-alias').Options,
-): import('lerna-alias').Aliases;
-
-export const getDefaultEnvironmentVars: (
-  isProduction?: boolean | undefined,
-  otherValues?: Record<string, string> | undefined,
-) => {
-  'process.env.NODE_ENV': string;
-};
-
-export const workspaceRoot: string;
-
-export declare const eslintConstants: {
-  /** List of file extensions to lint, in format used by eslint (comma-separated, with leading .) */
-  extensions: string;
-
-  /** Subdirectory to lint within package (relative path) */
-  directory: string;
-};
+export type { PackageJson, AllPackageInfo } from './types';
+export { default as getDependencies } from './getDependencies';
+export { default as findGitRoot } from './findGitRoot';
+export { default as findRepoDeps } from './findRepoDeps';
+export { default as getAllPackageInfo } from './getAllPackageInfo';
+export { default as isConvergedPackage } from './isConvergedPackage';
+export { default as getAffectedPackages } from './getAffectedPackages';
+export { getLernaAliases } from './get-lerna-aliases';
+export { getDefaultEnvironmentVars } from './getDefaultEnvironmentVars';
+export { getProjectMetadata, workspaceRoot } from './utils';
+export * as eslintConstants from './eslint-constants';
