+++ scripts/monorepo/types.ts
@@ -1,0 +1,21 @@
+export interface PackageJson {
+  name: string;
+  version: string;
+  private?: boolean;
+  main: string;
+  types?: string;
+  module?: string;
+  dependencies?: { [key: string]: string };
+  devDependencies?: { [key: string]: string };
+  peerDependencies?: { [key: string]: string };
+  [key: string]: any;
+}
+
+export interface PackageInfo {
+  /** Package path relative to the project root */
+  packagePath: string;
+  /** package.json contents */
+  packageJson: PackageJson;
+}
+
+export type AllPackageInfo = { [packageName: string]: PackageInfo };
